requirements.txt
psycopg[binary]>=3.1
PyYAML>=6.0
requests>=2.31
